Add user registraion and user roles

+++ src/store/modules/product.module.js
@@ -1,4 +1,5 @@
 import axios from '@/axios/firebase';
+import router from '@/router';
 
 export default {
   namespaced: true,
@@ -36,7 +37,6 @@ export default {
   actions: {
     async getProductList({commit}) {
       const {data} = await axios.get('/products.json')
-      // firebase возвращает массив с уникальными id, приходится парсить
       const unsortedProducts = Object.keys(data).map(key => {
         return {id: key, ...data[key]}
       })
@@ -49,32 +49,64 @@ export default {
     },
     async getCategoriesList({commit}) {
       const {data} = await axios.get('/categories.json')
-      // firebase возвращает массив с уникальными id, приходится парсить
       const categories = Object.keys(data).map(key => {
         return {id: key, ...data[key]}
       })
       
       commit('setCategories', categories)
     },
-    async create({commit, getters}, product) {
+    async create({commit, dispatch}, product) {
       const {data} = await axios.post('/products.json', product)
-      commit('addProduct', {...product, id: data.id})
+      commit('addProduct', {...product, id: data.name})
+      dispatch('setMessage', {
+        value: `Продукт ${product.title} добавлен`,
+        type: 'primary'
+      }, {root: true })
     },
-    async createCategory({commit, getters}, category) {
+    async createCategory({commit, dispatch}, category) {
       const {data} = await axios.post('/categories.json', category)
-      commit('addCategories', {...category, id: data.id})
+      commit('addCategory', {...category, id: data.id})
+      dispatch('setMessage', {
+        value: `Категория ${category.title} добавлена`,
+        type: 'primary'
+      }, {root: true })
     },
-    async remove({commit, getters}, id) {
+    async remove({commit, dispatch, getters}, id) {
+      const productTitle = getters.products.find(p => p.id === id).title
       await axios.delete(`/products/${id}.json`)
       commit('setProducts', getters.products.filter(p => p.id !== id))
+      await router.replace({path: '/admin/products'})
+      dispatch('setMessage', {
+        value: `Продукт ${productTitle} удалён`,
+        type: 'primary'
+      }, {root: true })
     },
-    async removeCategory({commit, getters}, id) {
-      await axios.delete(`/categories/${id}.json`)
-      commit('setCategories', getters.categories.filter(p => p.id !== id))
+    async removeCategory({commit, dispatch, getters}, id) {
+      const category = getters.categories.find(c => c.id === id)
+      if (getters.products.find(p => p.category === category.type)) {
+        dispatch('setMessage', {
+          value: `Категория ${category.title} не пустая`,
+          type: 'danger'
+        }, {root: true})
+      } else {
+        await axios.delete(`/categories/${id}.json`)
+        commit('setCategories', getters.categories.filter(p => p.id !== id))
+        await router.replace({path: '/admin/products'})
+        
+        dispatch('setMessage', {
+          value: `Категория ${category.title} удалена.`,
+          type: 'primary'
+        }, {root: true})
+      }
     },
-    async update({commit, getters}, product) {
+    async update({commit, dispatch}, product) {
       const {data} = await axios.put(`/products/${product.id}.json`, product)
-      console.log(data)
+      dispatch('setMessage', {
+        value: `Продукт ${data.title} обновлён`,
+        type: 'primary'
+      }, {root: true})
+      
+      return data
     }
   }
 }
